# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is program I developed as a lerning process before creating an API in Golang.",
        "title": "Easy API with Go",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/tcnksm/tool/blob/master/LICENCE"
        },
        "version": "1.0.0"
    },
    "host": "http://localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/user/{name}": {
            "get": {
                "description": "Display the specified person.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a personal user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Users": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}